Plotted to window
Worked on the plotting aspect of the code. Plotted functions to a new window. Commented about how we can get the input from the textbox. We can decide to use a button on the window or use a key press like "Enter".  After we receive that action than we can set! the functions and plot again.